refactor: extract sub-widgets from active_timer_screen
Split 564-line screen into four focused files under widgets/:
RingPainter, TimerGlassButton, TimerProgressRing, TimerPhaseIndicator.
